test: merge the two test_particle.py
